cli/azd/extensions/azure.foundry.ai.agents/tests/minimalHobo/src/echo-agent/requirements.txt
# PuPr https://github.com/microsoft/agent-framework/tree/main/python
agent-framework==1.0.0b251001

# Agents Hosting Adapter (local wheel due to PrPr)
./wheels/azure_ai_agentshosting-0.0.1+g9cac45e.d251002-py3-none-any.whl
# -e git+https://github.com/microsoft/container_agents.git#egg=azure-ai-agentshosting&subdirectory=src/adapter/python

pytest==8.4.2
azure-identity==1.25.0
python-dotenv==1.1.1
